requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
pillow>=10.2.0
opencv-python-headless>=4.9.0.80
python-magic>=0.4.27
aiofiles>=23.2.1
jinja2>=3.1.3
apscheduler>=3.10.4
